requirements.txt
# Core dependencies
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Database
# SQLite is included in Python standard library

# File watching
watchdog>=3.0.0

# Web-UI
flask>=3.0.0
werkzeug>=3.0.0
waitress>=2.1.2

# Configuration
PyYAML>=6.0.1

# CLI
# argparse is included in Python standard library

# Utilities
python-dateutil>=2.8.2
